Added fast gaussian very fast path for neon

+++ src/concurrency.hpp
@@ -147,12 +147,12 @@ void parallel_for_with_thread_id(const int numThreads, const int numIterations, 
 }
 
 template<typename Function, typename... Args>
-void parallel_for_segment(const int numThreads, const int numIterations, Function &&func, Args &&... args) {
+void parallel_for_segment(const int numThreads, const uint32_t numIterations, Function &&func, Args &&... args) {
   static_assert(std::is_invocable_v<Function, int, int, Args...>, "func must take an int parameter for iteration id");
 #if THREADS_SUPPORTED
   std::vector<std::thread> threads;
 
-  int segmentHeight = numIterations / numThreads;
+  int segmentHeight = numIterations / static_cast<uint32_t >(numThreads);
 
   auto parallelWorker = [&](int start, int end) {
     std::invoke(func, start, end, std::forward<Args>(args)...);
@@ -161,8 +161,8 @@ void parallel_for_segment(const int numThreads, const int numIterations, Functio
   if (numThreads > 1) {
     // Launch N-1 worker threads
     for (int i = 1; i < numThreads; ++i) {
-      int start = i * segmentHeight;
-      int end = (i + 1) * segmentHeight;
+      uint32_t start = i * segmentHeight;
+      uint32_t end = (i + 1) * segmentHeight;
       if (i == numThreads - 1) {
         end = numIterations;
       }
